Add tasks for db creation and seeding
The NoBrainer ORM does not correctly create database primary
or secondary indexes for fields that use an aliased name in the
database, so these must be defined manually. This commit adds
the rake task db:create which will create the required tables
along with the correct indexes. In addition, the db/seed file
now contains a subset of anonymised production data that can
be used for testing.