Remove empty production entries from application.yml
I think these empty entries were causing the environment variables
to be set to nil when running in production. To confirm.
Also removed Figaro initializer, which threw an exception if
a required env var was not set. These can't exist on the server
on the first deploy (they can only be set afterwards with dokku)
so the raised exception would abort the deployment.