Refactor js form_view.send method. Fixes #54. Fixes #56.
Issue #56 (Submitting form without an image raises Dragonfly exception)
caused by nil reference for an Image being cast to string 'undefined'
when constructing the Backbone FormData object to send to the server. Fixed
by making an explicit check for the existance of the image (and all
other attributes) before appending them to the form object, so the
attribute doesn't exist in the payload and is handled correctly on
the server.